+++ Packages/com.unity.inputsystem/Tests/TestFixture/InputTestFixture.cs
@@ -166,6 +166,10 @@ namespace UnityEngine.InputSystem
 
             try
             {
+#if UNITY_EDITOR
+                AssetDatabaseUtils.Restore();
+#endif
+
 #if DEVELOPMENT_BUILD || UNITY_EDITOR
                 InputSystem.Restore();
 #endif